Rename platforms to lib-i.
