+++ inputs/devices/base.py
@@ -5,10 +5,10 @@ import os
 import io
 
 
-from ..platforms.c import EVENT_SIZE, iter_unpack
-from ..errors import NoDevicePath, PERMISSIONS_ERROR_TEXT
-from ..platforms.system import NIX, WIN, MAC
-from ..platforms.event import InputEvent
+from ..libi.c import EVENT_SIZE, iter_unpack
+from ..libi.errors import NoDevicePath, PERMISSIONS_ERROR_TEXT
+from ..libi.system import NIX, WIN, MAC
+from ..libi.event import InputEvent
 
 
 class InputDevice(object):  # pylint: disable=useless-object-inheritance
