+++ inputs/devices/gamepad/gamepad.py
@@ -19,9 +19,9 @@ from operator import itemgetter
 import time
 
 from ...constants import XINPUT_ERROR_DEVICE_NOT_CONNECTED, XINPUT_ERROR_SUCCESS
-from ...errors import PERMISSIONS_ERROR_TEXT, UnknownEventType, UnpluggedError
-from ...platforms.c import (EVENT_FORMAT, convert_timeval)
-from ...platforms.system import NIX, WIN
+from ...libi.errors import PERMISSIONS_ERROR_TEXT, UnknownEventType, UnpluggedError
+from ...libi.c import (EVENT_FORMAT, convert_timeval)
+from ...libi.system import NIX, WIN
 from ..base import InputDevice
 from ._win import XinputGamepad, XinputState, XinputVibration, delay_and_stop
