+++ inputs/devices/keyboard/keyboard.py
@@ -1,7 +1,7 @@
 """Keyboard device class."""
 
 from ...constants import APPKIT_KB_PATH
-from ...platforms.system import NIX, WIN, MAC
+from ...libi.system import NIX, WIN, MAC
 from ..base import InputDevice
 from ._mac import mac_keyboard_process
 from ._win import win_keyboard_process
