+++ inputs/devices/mouse/mouse.py
@@ -1,7 +1,7 @@
 """Mouse input device handling."""
 
 from ...constants import APPKIT_MOUSE_PATH, QUARTZ_MOUSE_PATH
-from ...platforms.system import NIX, WIN, MAC
+from ...libi.system import NIX, WIN, MAC
 from ..base import InputDevice
 from ._win import win_mouse_process
 from ._mac import appkit_mouse_process, quartz_mouse_process
